Corrección de tamaño de texto Corrección conteo de votos, formato de las variables globales

+++ app/src/main/java/com/example/proyecto1zsjn/Usuario.java
@@ -4,6 +4,6 @@ import java.io.Serializable;
 
 public class Usuario{
 
-    public static int cand1, cand2, cand3;
+    public static float cand1, cand2, cand3;
 
 }
